- [x] (wait for another hightouch PR) Refactor rename and move the following to Actor space and out of Character space   - [x] (confirm) CharacterPortrait   - [x] (confirm) CharacterExhaustionBar   - [x] (confirm) src/sheets/quadrone/actor/character-parts/traits/CharacterTraitConfigurableListEntry.svelte   - [x] ~~(confirm) src/sheets/quadrone/actor/character-parts/traits/CharacterTraits.svelte~~ Only used by characters.   - [x] ~~(confirm) src/sheets/quadrone/actor/character-parts/CharacterSubtitle.svelte~~ Only used by characters.   - [x] ~~(confirm) src/sheets/quadrone/actor/character-parts/SavingThrowsCard.svelte~~ Only used by characters.
Added NPC punch list TODO items.
