+++ src/sheets/quadrone/actor/parts/ActorExhaustionBar.svelte
