+++ src/sheets/quadrone/actor/parts/ActorTraitConfigurableListEntry.svelte
